safari-600.7-branch/LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.5_Tokens/7.5.2_Keywords/S7.5.2_A1.1-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Unexpected token '='. Expected an identifier as the target for a break statement.
S7.5.2_A1.1

PASS Expected parsing failure

TEST COMPLETE
